(OraklNode) Update testcodes & fix migration file (#1145)
* feat: update testcodes, fix migration file

* fix: update based on feedback

+++ node/migrations/000001_initialize_tables.up.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS feeds (
     id SERIAL PRIMARY KEY,
     name TEXT NOT NULL UNIQUE,
     definition JSONB NOT NULL,
-    adapter_id INT16 NOT NULL,
+    adapter_id INT8 NOT NULL,
     CONSTRAINT feeds_adapter_id_fkey
         FOREIGN KEY(adapter_id)
         REFERENCES adapters(id)
